requirements.txt
streamlit>=1.24.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
fpdf2>=2.8.4
python-pptx>=0.6.21
openai>=1.0.0
